requirements.txt
pipecat-ai[deepgram]==1.0.0
websockets>=15.0
httpx>=0.27
python-dotenv>=1.0
